Moved components into components folder, improved scripts and contrib
