+++ src/components/Edge/Edge.types.ts
