+++ src/components/Node/Node.types.ts
