+++ src/components/Paper/index.ts
