+++ src/utilities/__tests__/workflowUtils.test.ts
@@ -1,4 +1,4 @@
-import { IPaperStoredNode } from '../../Paper';
+import { IPaperStoredNode } from '../../components/Paper';
 import {
   isNodeColliding,
   roundToNearest,
